requirements.txt
git+https://github.com/rapptz/discord-ext-menus
git+https://github.com/gorialis/jishaku
discord.py
numpydoc
cachetools
fuzzywuzzy
python-dotenv
parsedatetime
python-dateutil
humanize
asyncpg
asyncpg-stubs
lru-dict
aiofile
typing_extensions
python-Levenshtein
psutil
tabulate
aggdraw
colorthief
numpy
pydantic
matploblib
numpy

emoji